+++ desktop/cmp/contextmenu/ContextMenuItem.js
@@ -22,6 +22,7 @@ export class ContextMenuItem {
     items;
     disabled;
     hidden;
+    displayFn;
 
     /**
      * @param {Object} c - ContextMenuItem configuration.
@@ -32,6 +33,7 @@ export class ContextMenuItem {
      * @param {Object[]} [c.items] - child menu items.
      * @param {boolean} [c.disabled] - true to disable this item.
      * @param {boolean} [c.hidden] - true to hide this item.
+     * @param {DisplayFn} [c.displayFn] - called prior to showing this item in the UI.
      */
     constructor({
         text,
@@ -40,7 +42,8 @@ export class ContextMenuItem {
         actionFn = null,
         items = null,
         disabled = false,
-        hidden = false
+        hidden = false,
+        displayFn = null
     }) {
         this.text = text;
         this.icon = icon;
@@ -49,6 +52,7 @@ export class ContextMenuItem {
         this.items = items;
         this.disabled = disabled;
         this.hidden = hidden;
+        this.displayFn = displayFn;
     }
 
     /**
@@ -80,3 +84,8 @@ export class ContextMenuItem {
         }, defs);
     }
 }
+
+/**
+ * @callback DisplayFn - called prior to rendering the ContextMenuItem's UI element.
+ * @returns {Object} - display configs to override for this render of the item.
+ */